+++ src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/websocket/page/NodePath.kt
@@ -3,8 +3,8 @@ package com.tencent.devops.environment.websocket.page
 import com.tencent.devops.common.websocket.IPath
 import com.tencent.devops.common.websocket.pojo.BuildPageInfo
 
-class NodePath: IPath {
-	override fun buildPage(buildPageInfo: BuildPageInfo): String {
-		return "/console/environment/${buildPageInfo.projectId}/nodeList"
-	}
+class NodePath : IPath {
+    override fun buildPage(buildPageInfo: BuildPageInfo): String {
+        return "/console/environment/${buildPageInfo.projectId}/nodeList"
+    }
 }
