+++ src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/websocket/push/NodeWebsocketPush.kt
@@ -8,33 +8,32 @@ import com.tencent.devops.common.websocket.dispatch.message.SendMessage
 import com.tencent.devops.common.websocket.dispatch.push.WebsocketPush
 import com.tencent.devops.common.websocket.pojo.NotifyPost
 import com.tencent.devops.common.websocket.pojo.WebSocketType
-import com.tencent.devops.environment.websocket.NodeMessage
+import com.tencent.devops.common.websocket.dispatch.message.NodeMessage
 
 @Event(exchange = MQ.EXCHANGE_WEBSOCKET_TMP_FANOUT, routeKey = MQ.ROUTE_WEBSOCKET_TMP_EVENT)
 data class NodeWebsocketPush(
-		val projectId: String,
-		override val userId: String,
-		override val pushType: WebSocketType,
-		override val redisOperation: RedisOperation,
-		override val objectMapper: ObjectMapper,
-		override var page: String?,
-		override var notifyPost: NotifyPost
-): WebsocketPush(userId, pushType, redisOperation, objectMapper, page, notifyPost) {
-	override fun findSession(page: String): List<String>? {
-		return super.findSession(page)
-	}
+    val projectId: String,
+    override val userId: String,
+    override val pushType: WebSocketType,
+    override val redisOperation: RedisOperation,
+    override val objectMapper: ObjectMapper,
+    override var page: String?,
+    override var notifyPost: NotifyPost
+) : WebsocketPush(userId, pushType, redisOperation, objectMapper, page, notifyPost) {
+    override fun findSession(page: String): List<String>? {
+        return super.findSession(page)
+    }
 
-	override fun buildMqMessage(): SendMessage? {
-		return NodeMessage(
-				project = projectId,
-				userId =  userId,
-				page = page,
-				sessionList = findSession(page!!),
-				notifyPost = notifyPost
-		)
-	}
+    override fun buildMqMessage(): SendMessage? {
+        return NodeMessage(
+                project = projectId,
+                userId = userId,
+                page = page,
+                sessionList = findSession(page!!),
+                notifyPost = notifyPost
+        )
+    }
 
-	override fun buildNotifyMessage(message: SendMessage) {
-
-	}
+    override fun buildNotifyMessage(message: SendMessage) {
+    }
 }
